+++ src/article/editor/EditorPackage.js
@@ -50,7 +50,7 @@ import InsertDispQuoteCommand from './InsertDispQuoteCommand'
 import InsertExtLinkCommand from './InsertExtLinkCommand'
 import InsertFigureCommand from './InsertFigureCommand'
 import InsertFigureTool from './InsertFigureTool'
-import { InsertFigurePanelTool, ReplaceFigurePanelTool } from '../shared/FigurePanelTools'
+import InsertFigurePanelTool from '../shared/InsertFigurePanelTool'
 import InsertFootnoteCommand from '../shared/InsertFootnoteCommand'
 import InsertFootnoteCrossReferenceCommand from './InsertFootnoteCrossReferenceCommand'
 import InsertInlineFormulaCommand from './InsertInlineFormulaCommand'
@@ -64,6 +64,7 @@ import {
 } from './TableCommands'
 import InsertTableTool from './InsertTableTool'
 import RemoveItemCommand from './RemoveItemCommand'
+import ReplaceFigurePanelTool from '../shared/ReplaceFigurePanelTool'
 import ReplaceSupplementaryFileCommand from './ReplaceSupplementaryFileCommand'
 import ReplaceSupplementaryFileTool from './ReplaceSupplementaryFileTool'
 import {
