+++ src/article/shared/InsertFigurePanelTool.js
@@ -1,0 +1,7 @@
+import UploadSingleImageTool from './UploadSingleImageTool'
+
+export default class InsertFigurePanelTool extends UploadSingleImageTool {
+  getClassNames () {
+    return 'sc-insert-figure-panel-tool sc-upload-tool sc-tool'
+  }
+}
