+++ src/article/shared/ReplaceFigurePanelTool.js
@@ -1,0 +1,7 @@
+import UploadSingleImageTool from './UploadSingleImageTool'
+
+export default class ReplaceFigurePanelTool extends UploadSingleImageTool {
+  getClassNames () {
+    return 'sc-replace-figure-panel-tool sc-upload-tool sc-tool'
+  }
+}
